% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/query.R
\name{article_pageviews}
\alias{article_pageviews}
\title{Retrieve Pageview Data for an Article}
\usage{
article_pageviews(project = "en.wikipedia", article = "Barack Obama",
  platform = "all", user_type = "all", granularity = "daily",
  start = "2015100100", end = NULL, ...)
}
\description{
retrieves the pageview data for a particular article on a project, within
a provided time-range.
}

