% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/urls.R
\name{extract_project}
\alias{extract_project}
\title{extract the project from Wikipedia URLs}
\usage{
extract_project(urls)
}
\arguments{
\item{urls}{a vector of URLs}
}
\value{
a vector of extracted projects, in the form of project_variant.project_class.
}
\description{
take a vector of wikimedia URLs and extract the language or project variant,
along with the project class.
}

