% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/urls.R
\name{extract_referers}
\alias{extract_referers}
\title{normalise and extract referers}
\usage{
extract_referers(data)
}
\arguments{
\item{data}{a data.frame outputted by \code{\link{read_sampled_log}}.}
}
\value{
a data.frame matching the input in structure, but with normalised referers.
}
\description{
decodes, simplifies and summarises referers
in log lines, extracting high-level names (for example, "Google")
rather than inconsistently-fragmented URLs.
}

