% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/filter.R
\name{identify_access_method}
\alias{identify_access_method}
\title{identify the access method associated with a request}
\usage{
identify_access_method(urls)
}
\arguments{
\item{urls}{a vector of URLs from \code{\link{read_sampled_log}}.}
}
\description{
identify whether the request is via desktop, the mobile web
or the mobile app.
}

