% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/filter.R
\name{is_spider}
\alias{is_spider}
\title{check whether a request is from a spider}
\usage{
is_spider(user_agents)
}
\arguments{
\item{user_agents}{a vector of user agents, which can be retrieved with
\code{\link{read_sampled_log}}.}
}
\value{
a boolean vector identifying whether the user agent at the equivalent indices
in the input vector matched that of a spider/web crawler or not.
}
\description{
consumes a vector of user agents and checks whether each one
matches the ua-parser spider definitions. In addition, it looks for Wikimedia-specific
spiders that aren't included in the (generalised) spider list.
}

