% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/package.R
\docType{package}
\name{ pageviews }
\alias{ pageviews }
\alias{ pageviews }
\title{ An API client for Wikimedia traffic data }
\description{
Pageview data from the 'Wikimedia' sites, such as Wikipedia (https://www.wikipedia.org/), from entire projects to by-article levels of granularity.
}
