% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/pageviews.R
\docType{package}
\name{pageviews}
\alias{pageviews}
\alias{pageviews-package}
\title{Functions for handling pageviews data}
\description{
This package provides functions for handling the Wikimedia request logs
and performing a variety of operations on them, most prominently extracting "pageviews"
and classifying them.
}

