% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/query.R
\name{project_pageviews}
\alias{project_pageviews}
\title{Retrieve Per-Project Pageview Counts}
\usage{
project_pageviews(project = "en.wikipedia", platform = "all",
  user_type = "all", granularity = "daily", start = "2015100100",
  end = NULL, reformat = TRUE, ...)
}
\description{
Retrieve Per-Project Pageview Counts
}

