% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/read.R
\name{read_sampled_log}
\alias{read_sampled_log}
\title{read a sampled log file}
\usage{
read_sampled_log(file)
}
\arguments{
\item{file}{a filename, retrieved with \code{\link{get_files}}}
}
\value{
a data.frame containing 16 columns - "squid", "sequence_no", "timestamp",
"servicetime", "ip_address", "status_code", "reply_size", "request_method",
"url", "squid_status", "mime_type", "referer", "x_forwarded", "user_agent",
"lang" and "x_analytics".
}
\description{
read a sampled log file identified with \code{\link{get_files}}.
The sampled logs are returned as a data.frame with 16 columns - see
the "Value" documentation.
}
\seealso{
\code{\link{to_pageviews}}, for filtering the output
}

