% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/query.R
\name{top_articles}
\alias{top_articles}
\title{Retrieve Data on Top Articles}
\usage{
top_articles(project = "en.wikipedia", platform = "all", year = "all",
  month = "all", day = "all", ...)
}
\arguments{
\item{project}{the name of the project, structured as \code{[language_code].[project]}
(see the default).}

\item{platform}{The platform the pageviews came from; one of "all", "desktop", "mobile-web" and
"mobile-app". Set to "all" by default.}

\item{year}{The year the articles were "top" in. "all" by default; if it is "all", month
and day must also be "all".}

\item{month}{The month the articles were "top" in. "all" by default; if it is "all", day must also
be "all".}

\item{day}{The day the articles were "top" in. "all" by default; this simply returns the top articles
for the relevant month/year combination.}
}
\description{
\code{top_articles} grabs data on the top articles for a project
in a given time period, and for a particular platform.
}
\seealso{
\code{\link{article_pageviews}} for per-article pageviews and \code{\link{project_pageviews}} for
per-project pageviews.
}

