% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pageviews.R
\docType{package}
\name{pageviews}
\alias{pageviews}
\title{An API client for Wikimedia traffic data}
\description{
Pageview data from the 'Wikimedia' sites, such as Wikipedia (https://www.wikipedia.org/),
from entire projects to by-article levels of granularity.
}
